requirements-dev.txt
black
coverage==7.4.0
django-browser-reload==1.11.0
django-coverage-plugin==3.1.0
django-debug-toolbar==4.2.0
django-extensions==3.2.3
djlint==1.34.1
factory-boy==3.3.0
pip-tools==7.3.0
pre-commit
pytest-cov==4.1.0
pytest-factoryboy==2.6.0
pytest-django==4.7.0
